MLOps/hw2/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.62.1
joblib==1.5.2
numpy==2.3.5
protobuf==6.33.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
